vlm_service/requirements.txt
fastapi
uvicorn
python-multipart
requests
Pillow
pydantic
torch
torchvision
transformers
accelerate
qwen_vl_utils
easyocr
paddlepaddle
paddleocr
supervision==0.18.0
ultralytics==8.3.70
numpy==1.26.4
opencv-python-headless
matplotlib
